Added Cloudinary storage + Dotenv
All the images of campgrounds are now hosted from cloud and at the same time, they are being saved in the cloudinary storage. The images are added/deleted to both mongodb and cloudinary storage. Major changes are added to the UI of new.ejs and edit.ejs as well. While revising, look closely how the cloudinary is configured.